+++ web/src/features/fba/pages/FireBehaviourAdvisoryPage.tsx
@@ -1,9 +1,9 @@
 import { FormControl, Grid, makeStyles } from '@material-ui/core'
 import { GeneralHeader, Container } from 'components'
 import React, { useEffect, useState } from 'react'
-import FBAMap from 'features/fbaCalculator/components/map/FBAMap'
+import FBAMap from 'features/fba/components/FBAMap'
 import FireCenterDropdown from 'features/fbaCalculator/components/FireCenterDropdown'
-import FormalFBATable from 'features/fbaCalculator/components/FormalFBATable'
+import FormalFBATable from 'features/fba/components/FormalFBATable'
 import DatePicker from 'components/DatePicker'
 import { DateTime } from 'luxon'
 import { selectFireCenters } from 'app/rootReducer'
